+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+### Removed
+
+- Dependency `tarampampam/wrappers-php` because this package was deprecated and removed
+
+### Changed
+
+- Minimal `PHP` version now is `^7.3`
+
 ## v2.4.0
 
 ### Added